sections/10_langgraph_introduction/projects/react_agent_with_langgraph/pyproject.toml
[project]
name = "react-agent-with-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-tavily>=0.2.13",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.13",
]
